// src/services/types/filter-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {FilterItemOptionResponse} from "./filter-item-option-response.ts";

export type FilterResponse = {
	/**
	 * @description 필터 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 테이블명
	 * @type string
	 */
	table: string;
	/**
	 * @description 컬럼명
	 * @type string
	 */
	field: string;
	/**
	 * @description 컬럼 레이블
	 * @type string
	 */
	label: string;
	/**
	 * @description 컬럼 타입
	 * @type string
	 */
	dataType: string;
	/**
	 * @description 연산자
	 * @type string
	 */
	operator: string;
	/**
	 * @description 입력 컴포넌트
	 * @type string
	 */
	component: string;
	/**
	 * @description 컴포넌트 요소
	 * @type array | undefined
	 */
	items?: FilterItemOptionResponse[];
	/**
	 * @description 생성자 아이디
	 * @type integer, int64
	 */
	creatorId: number;
	/**
	 * @description 생성 일시
	 * @type string, date-time
	 */
	createdAt: string;
	/**
	 * @description 수정 일시
	 * @type string | undefined, date-time
	 */
	updatedAt?: string;
};