// src/services/types/filter-item-option-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type FilterItemOptionResponse = {
	/**
	 * @description 노출 라벨
	 * @type string
	 */
	label: string;
	/**
	 * @description 실제 값
	 * @type string
	 */
	value: string;
};