// src/services/types/media-library/media-library-search-media-library.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {ApiResultOnePageMapStringObject} from "../api-result-one-page-map-string-object.ts";

export type MediaLibrarySearchMediaLibraryQueryParams = {
	/**
	 * @description private 필드 포함 여부
	 * @default false
	 * @type boolean | undefined
	 */
	includePrivate?: boolean;
	/**
	 * @description 조회 대상 필드 목록
	 * @type array | undefined
	 */
	field?: string[];
	/**
	 * @description 조회 연산자 목록
	 * @type array | undefined
	 */
	operator?: string[];
	/**
	 * @description 조회 값 목록
	 * @type array | undefined
	 */
	value?: string[];
	/**
	 * @type integer | undefined, int32
	 */
	page?: number;
	/**
	 * @type integer | undefined, int32
	 */
	size?: number;
};

/**
 * @description OK
 */
export type MediaLibrarySearchMediaLibrary200 = ApiResultOnePageMapStringObject;

export type MediaLibrarySearchMediaLibraryQueryResponse = MediaLibrarySearchMediaLibrary200;

export type MediaLibrarySearchMediaLibraryQuery = {
	Response: MediaLibrarySearchMediaLibrary200;
	QueryParams: MediaLibrarySearchMediaLibraryQueryParams;
	Errors: any;
};